fix: fixing some bugs

+++ Services/MatchService.cs
@@ -22,29 +22,29 @@ public class MatchService
         _goalService = goalService;
     }
 
-    public async Task EndGameToKnockoutValidationAsync(int matchId)
+    public async Task EndGameToKnockoutValidationAsync(int matchId, bool valid)
     {
         var match = await GetMatchById(matchId);
-        if(match is null)
+        if(match is null && valid)
         {
             throw new ApplicationException("Partida passada não existe.");
         }
-        if(await DepartureDateNotSet(matchId))
+        if(await DepartureDateNotSet(matchId) && valid)
         {
             throw new ApplicationException("Data da partida não definida.");
         }
-        if(match.Date.ToUniversalTime() >= DateTime.UtcNow)
+        if(match.Date.ToUniversalTime() >= DateTime.UtcNow && valid)
         {
             throw new ApplicationException("Partida ainda não inciou");
         }
-        if(match.Winner != 0)
+        if(match.Winner != 0 && valid)
         {
             throw new ApplicationException("Partida já possui um vencedor.");
         }
 
         var championship = await GetByIdSend(match.ChampionshipId);
 
-        if(match.HomeUniform is null || match.VisitorUniform is null)
+        if((match.HomeUniform is null || match.VisitorUniform is null) && valid)
             throw new ApplicationException("É necessário definir os uniformes das equipes antes");
         
         if (match.Road is null)
@@ -55,7 +55,7 @@ public class MatchService
 		{
             var aggregateVisitorPoints = await GetPointsFromTeamByIdInTwoMatches(matchId, match.Visitor);
             var aggregateHomePoints = await GetPointsFromTeamByIdInTwoMatches(matchId, match.Home);
-            if(match.PreviousMatch != 0)
+            if(match.PreviousMatch != 0 && valid)
             {
                 if(aggregateVisitorPoints == aggregateHomePoints && match.Winner == 0)
                 {
@@ -612,37 +612,37 @@ public class MatchService
     private async Task<bool> CheckIfThereIsWinner(int matchId)
         => await _dbService.GetAsync<bool>("SELECT EXISTS(SELECT * FROM matches WHERE id = @matchId AND Winner IS NOT NULL)", new {matchId});
 
-    public async Task EndGameToLeagueSystemValidationAsync(int matchId)
+    public async Task EndGameToLeagueSystemValidationAsync(int matchId, bool valid)
     {
         var match = await GetMatchById(matchId);
         var championship = await GetChampionshipByMatchId(matchId);
-        // if(match is null)
-        // {
-        //     throw new ApplicationException("Partida passada não existe.");
-        // }
-        // if(await DepartureDateNotSet(matchId))
-        // {
-        //     throw new ApplicationException("Data da partida não definida.");
-        // }
-        // if(match.Date.ToUniversalTime() >= DateTime.UtcNow)
-        // {
-        //     throw new ApplicationException("Partida ainda não inciou");
-        // }
-        // if(match.Winner != 0)
-        // {
-        //     throw new ApplicationException("Partida já possui um vencedor.");
-        // }
-        // if(match.Tied)
-        // {
-        //     throw new ApplicationException("Partida já terminou em empate.");
-        // }
-        // if(match.HomeUniform is null || match.VisitorUniform is null)
-        //     throw new ApplicationException("É necessário definir os uniformes das equipes antes");
+        if(match is null && valid)
+        {
+            throw new ApplicationException("Partida passada não existe.");
+        }
+        if(await DepartureDateNotSet(matchId) && valid)
+        {
+            throw new ApplicationException("Data da partida não definida.");
+        }
+        if(match.Date.ToUniversalTime() >= DateTime.UtcNow && valid)
+        {
+            throw new ApplicationException("Partida ainda não inciou");
+        }
+        if(match.Winner != 0 && valid)
+        {
+            throw new ApplicationException("Partida já possui um vencedor.");
+        }
+        if(match.Tied && valid)
+        {
+            throw new ApplicationException("Partida já terminou em empate.");
+        }
+        if((match.HomeUniform is null || match.VisitorUniform is null) && valid)
+            throw new ApplicationException("É necessário definir os uniformes das equipes antes");
         
-        // if(match.Road is null)
-        //     throw new ApplicationException("É necessário definir o local da partida antes");
-        // if(await CheckIfLastMatchHasEnded(match))
-        //     throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
+        if(match.Road is null && valid)
+            throw new ApplicationException("É necessário definir o local da partida antes");
+        if(await CheckIfLastMatchHasEnded(match) && valid)
+            throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
 
         var visitorPoints = await GetPointsFromTeamById(matchId, match.Visitor);
         var homePoints = await GetPointsFromTeamById(matchId, match.Home);
@@ -1104,39 +1104,39 @@ public class MatchService
         return tempCards + userCards;
     }
 
-    public async Task EndGameToGroupStageValidationAsync(int matchId)
+    public async Task EndGameToGroupStageValidationAsync(int matchId, bool valid)
     {
         var match = await GetMatchById(matchId);
         var championship = await GetChampionshipByMatchId(matchId);
-        // if(match is null)
-        // {
-        //     throw new ApplicationException("Partida passada não existe.");
-        // }
-        // if(await DepartureDateNotSet(matchId))
-        // {
-        //     throw new ApplicationException("Data da partida não definida.");
-        // }
-        // if(match.Date.ToUniversalTime() >= DateTime.UtcNow)
-        // {
-        //     throw new ApplicationException("Partida ainda não inciou");
-        // }
-        // if(match.Winner != 0)
-        // {
-        //     throw new ApplicationException("Partida já possui um vencedor.");
-        // }
-        // if(match.Tied)
-        // {
-        //     throw new ApplicationException("Partida já terminou em empate.");
-        // }
+        if(match is null && valid)
+        {
+            throw new ApplicationException("Partida passada não existe.");
+        }
+        if(await DepartureDateNotSet(matchId) && valid)
+        {
+            throw new ApplicationException("Data da partida não definida.");
+        }
+        if(match.Date.ToUniversalTime() >= DateTime.UtcNow && valid)
+        {
+            throw new ApplicationException("Partida ainda não inciou");
+        }
+        if(match.Winner != 0 && valid)
+        {
+            throw new ApplicationException("Partida já possui um vencedor.");
+        }
+        if(match.Tied && valid)
+        {
+            throw new ApplicationException("Partida já terminou em empate.");
+        }
 
-        // if(match.HomeUniform is null || match.VisitorUniform is null)
-        //     throw new ApplicationException("É necessário definir os uniformes das equipes antes");
+        if((match.HomeUniform is null || match.VisitorUniform is null) && valid)
+            throw new ApplicationException("É necessário definir os uniformes das equipes antes");
         
-        // if (match.Road is null)
-        //     throw new ApplicationException("É necessário definir o local da partida antes antes");
+        if (match.Road is null && valid)
+            throw new ApplicationException("É necessário definir o local da partida antes antes");
         
-        // if(await CheckIfLastMatchHasEnded(match))
-        //     throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
+        if(await CheckIfLastMatchHasEnded(match) && valid)
+            throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
 
         var visitorPoints = await GetPointsFromTeamById(matchId, match.Visitor);
         var homePoints = await GetPointsFromTeamById(matchId, match.Home);
@@ -1936,32 +1936,32 @@ public class MatchService
     private async Task<User> GetUserById(Guid id)
         => await _dbService.GetAsync<User>("SELECT * FROM Users WHERE Id = @id", new {id});
 
-    public async Task WoValidation(int matchId, int teamId)
+    public async Task WoValidation(int matchId, int teamId, bool valid)
     {
         var match = await GetMatchById(matchId);
         var championship = await GetChampionshipByMatchId(matchId);
 
         var player = await GetPlayerOfteamSend(teamId != match.Visitor ? match.Visitor : match.Home );
 
-        // if(match is null)
-        //     throw new ApplicationException("Partida passada não existe");
-        // if(await DepartureDateNotSet(matchId))
-        //     throw new ApplicationException("Data da partida não definida.");
-        // if(match.Date.ToUniversalTime() >= DateTime.UtcNow)
-        //     throw new ApplicationException("Partida ainda não inciou");
-        // if(match.Winner != 0)
-        //     throw new ApplicationException("Partida já possui um vencedor.");
-        // if(match.HomeUniform is null || match.VisitorUniform is null)
-        //     throw new ApplicationException("É necessário definir os uniformes das equipes antes");
-        // if (match.Road is null)
-        //     throw new ApplicationException("É necessário definir o local da partida antes antes");
+        if(match is null && valid)
+            throw new ApplicationException("Partida passada não existe");
+        if(await DepartureDateNotSet(matchId) && valid)
+            throw new ApplicationException("Data da partida não definida.");
+        if(match.Date.ToUniversalTime() >= DateTime.UtcNow && valid)
+            throw new ApplicationException("Partida ainda não inciou");
+        if(match.Winner != 0 && valid)
+            throw new ApplicationException("Partida já possui um vencedor.");
+        if((match.HomeUniform is null || match.VisitorUniform is null) && valid)
+            throw new ApplicationException("É necessário definir os uniformes das equipes antes");
+        if (match.Road is null && valid)
+            throw new ApplicationException("É necessário definir o local da partida antes antes");
 
         if(match.Round != 0 && championship.Format == Format.GroupStage)
         {
-            // if(match.Tied)
-            //     throw new ApplicationException("Partida já terminou em empate.");
-            // if(await CheckIfLastMatchHasEnded(match))
-            //     throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
+            if(match.Tied && valid)
+                throw new ApplicationException("Partida já terminou em empate.");
+            if(await CheckIfLastMatchHasEnded(match) && valid)
+                throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
             
             await _goalService.RemoveAllGoalOfMatchValidation(match.Id);
 
@@ -1993,14 +1993,14 @@ public class MatchService
                 }
             }
             
-            await EndGameToGroupStageValidationAsync(match.Id);
+            await EndGameToGroupStageValidationAsync(match.Id, false);
         }
         else if(match.Round != 0 && championship.Format == Format.LeagueSystem)
         {
-            // if(match.Tied)
-            //     throw new ApplicationException("Partida já terminou em empate.");
-            // if(await CheckIfLastMatchHasEnded(match))
-            //     throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
+            if(match.Tied && valid)
+                throw new ApplicationException("Partida já terminou em empate.");
+            if(await CheckIfLastMatchHasEnded(match) && valid)
+                throw new ApplicationException("Partida anterior de um dos times anda não foi finalizada");
             
             await _goalService.RemoveAllGoalOfMatchValidation(match.Id);
 
@@ -2032,7 +2032,7 @@ public class MatchService
                 }
             }
 
-            await EndGameToLeagueSystemValidationAsync(match.Id);
+            await EndGameToLeagueSystemValidationAsync(match.Id, false);
         }
 
         else
@@ -2066,7 +2066,7 @@ public class MatchService
                     await CreateGoalToPlayerSend(goal);
                 }
             }
-            await EndGameToKnockoutValidationAsync(matchId);
+            await EndGameToKnockoutValidationAsync(matchId, false);
         }
     }
     private async Task<User> GetPlayerOfteamSend(int id) =>
